requirements.txt
requests==2.28.0
dotmap==1.3.30
PyJWT==2.4.0